+++ src/pint/models/timing_model.py
@@ -9,6 +9,7 @@ import copy
 import inspect
 from collections import OrderedDict, defaultdict
 from functools import wraps
+from warnings import warn
 
 import astropy.time as time
 import astropy.units as u
@@ -21,6 +22,8 @@ import pint
 from pint.models.parameter import (
     AngleParameter,
     MJDParameter,
+    boolParameter,
+    floatParameter,
     Parameter,
     maskParameter,
     strParameter,
@@ -40,14 +43,6 @@ __all__ = ["DEFAULT_ORDER", "TimingModel"]
 # Comparisons with keywords in par file lines is done in a case insensitive way.
 ignore_params = set(
     [
-        "START",
-        "FINISH",
-        "EPHVER",
-        "UNITS",
-        "TIMEEPH",
-        "T2CMETHOD",
-        "DILATEFREQ",
-        "NTOA",
         "TRES",
         "TZRMJD",
         "TZRFRQ",
@@ -57,7 +52,6 @@ ignore_params = set(
         "BINARY",
         "CHI2R",
         "MODE",
-        "INFO",
         "PLANET_SHAPIRO2",
         #    'NE_SW', 'NE_SW2',
     ]
@@ -228,6 +222,62 @@ class TimingModel(object):
         self.add_param_from_top(
             MJDParameter(name="FINISH", description="End MJD for fitting"), ""
         )
+        self.add_param_from_top(
+            strParameter(
+                name="INFO",
+                description="Tells TEMPO to write some extra information about frontend/backend combinations; -f is recommended",
+            ),
+            "",
+        )
+        self.add_param_from_top(
+            strParameter(
+                name="TIMEEPH",
+                description="Time ephemeris to use for TDB conversion; for PINT, always FB90",
+            ),
+            "",
+        )
+        self.add_param_from_top(
+            strParameter(
+                name="T2CMETHOD",
+                description="Method for transforming from terrestrial to celestial frame (IAU2000B/TEMPO; PINT only supports ????)",
+            ),
+            "",
+        )
+        self.add_param_from_top(
+            boolParameter(
+                name="DILATEFREQ",
+                value=False,
+                description="Whether or not TEMPO2 should apply gravitational redshift and time dilation to obseerving frequency (Y/N; PINT only supports N)",
+            ),
+            "",
+        )
+        self.add_param_from_top(
+            floatParameter(
+                name="DMDATA",
+                value=0.0,
+                units="",
+                description="Was the fit done using per-TOA DM information?",
+            ),
+            "",
+        )
+        self.add_param_from_top(
+            floatParameter(
+                name="NTOA",
+                value=0.0,
+                units="",
+                description="Number of TOAs used in the fitting",
+            ),
+            "",
+        )
+        self.add_param_from_top(
+            floatParameter(
+                name="CHI2",
+                value=0.0,
+                units="",
+                description="Chi-squared value obtained during fitting",
+            ),
+            "",
+        )
 
         for cp in components:
             self.add_component(cp, validate=False)
@@ -246,6 +296,14 @@ class TimingModel(object):
 
         The checks include required parameters and parameter values.
         """
+        if self.DILATEFREQ.value:
+            warn("PINT does not support 'DILATEFREQ Y'")
+        if self.TIMEEPH.value not in [None, "FB90"]:
+            warn("PINT only supports 'TIMEEPH FB90'")
+        if self.T2CMETHOD.value not in [None, "IAU2000B"]:  # FIXME: really?
+            warn("PINT only supports 'T2CMETHOD IAU2000B'")
+        if self.UNITS.value not in [None, "TDB"]:
+            raise ValueError("PINT only supports 'UNITS TDB'")
         for cp in self.components.values():
             cp.validate()
 
@@ -1875,6 +1933,7 @@ class TimingModel(object):
         last_order=["jump_delay"],
     ):
         """Represent the entire model as a parfile string."""
+        self.validate()
         result_begin = ""
         result_end = ""
         result_middle = ""